add charge

+++ domain/src/main/java/com/paya/domain/models/repo/AddInventoryPriceRepoModel.kt
@@ -1,4 +1,4 @@
 package com.paya.domain.models.repo
 
 data class AddInventoryPriceRepoModel(val name: String,
-                                      var price: Float)
+                                      var price: Long)
